move workflow model to its own ressource
